requirements.txt
fastapi
uvicorn
streamlit
sqlalchemy
pydantic
python-multipart
python-jose
bcrypt==4.1.2
passlib[bcrypt]
pandas
scikit-learn==1.4.1.post1
numpy
matplotlib
seaborn
email-validator

# spaCy + English Model
spacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Transformers + Torch + Accelerate
transformers[torch]
accelerate==0.27.2
torch

requests
